[DuiBenKuN]
整理脚本。
